Fixed issue with using server defined DB connection in leases.go and leases_test.go.
